fix bad imports from lib/errors.ts
