+++ src/core/lib/yaml-schema/types.ts
@@ -1,14 +1,13 @@
 /*
-* types.ts
-*
-* Types for the YAML validator
-*
-* Copyright (C) 2022 Posit Software, PBC
-*
-*/
+ * types.ts
+ *
+ * Types for the YAML validator
+ *
+ * Copyright (C) 2022 Posit Software, PBC
+ */
 
 import { MappedString } from "../text-types.ts";
-import { ErrorLocation, TidyverseError } from "../errors.ts";
+import { ErrorLocation, TidyverseError } from "../errors-types.ts";
 
 // https://github.com/microsoft/TypeScript/issues/1897#issuecomment-822032151
 export type JSONValue =
